modify : song controller

+++ server/controllers/mypage/withdrawal.js
@@ -6,42 +6,41 @@ module.exports = async (req, res) => {
   try {
     // const accessTokenData = isAuthorized(req);
 
-    // if (!accessTokenData) {
-    //   return res.status(404).send({ message: 'You\'re not logged in.' });
-    // } else {
-    //   await User.destroy({
-    //     where: { id: accessTokenData.id },
-    //   });
-    // }
+    // JUST FOR TEST PURPOSES: without a real accessToken
+    // console.log(req.headers.authorization);
+    const accessTokenData = { id: req.headers.authorization };
 
-    // test: without accessToken
-    const userInfo = await user.findOne({
-      where: {
-        id: req.body.id
-      }
-    });
-
-    if (userInfo) {
-      await user.destroy({
+    if (!accessTokenData) {
+      return res.status(404).send({ message: 'You\'re not logged in.' });
+    } else {
+      const userInfo = await user.findOne({
         where: {
-          id: userInfo.id
+          id: accessTokenData.id
         }
       });
 
-      // res.setHeader('authorization', '');
-      // res.cookie('refreshToken', '');
+      if (userInfo) {
+        await user.destroy({
+          where: {
+            id: accessTokenData.id
+          }
+        });
 
-      res.status(200).json({
-        message: 'Successfully withdrawn'
-      });
-    } else {
-      res.status(400).json({
-        message: 'Invalid access token'
-      });
+        res.setHeader('authorization', '');
+        // res.cookie('refreshToken', '');
+
+        res.status(200).json({
+          message: 'Successfully withdrawn'
+        });
+      } else {
+        res.status(400).json({
+          message: 'Error'
+        });
+      }
     }
   } catch {
     res.status(400).json({
-      message: 'Invalid access token'
+      message: 'Error'
     });
   }
 };
